Validate 'user_date' parameter, and prevent compile-time bug
The user_date argument wasn't being validated, and there
was a possibility of a compile-time value being used for
initialization; I moved the calculation of initialization-
time date calculation into __init__ and used None as
the "omitted argument" sentinel.

I also used datetime.strptime to validate that the provided
argument is valid.