Separate except clauses for the different types of errors we can catch
Every different error type returns some context about the error that
caused it; also, never "swallow" errors silently; this function originally
returned None if there was an error, now it'll re-raise the same
exception, or it can be re-cast as a custom error type.

Also, execution does not proceed beyond the point when an error was
detected, simplifying the definition's structure.