Merge Develop into Master (#3)
* - Fixed typos
- Added `develop` branch
- Added `develop` to github workflow

* Suggestions and stuff (#2)

* - Fixed typos (#1)

- Added `develop` branch
- Added `develop` to github workflow

* Uniformized to the most frequently found format

* Changed from 'Exception' to 'ValueError' base exception

By itself 'Exception' is overbroad, and doesn't convey any
context about the error that triggered it.  ValueError is
expressly used for notifying on bad parameter or argument
values; also added 'Value' suffix, for more clarity.

* Delete '(object)' base class

Since the demise of Python 2.x there is no separation between
old- and new-style classes, so there is no need to inherit
from 'object' anymore; it's considered superfluous and redundant
and redundant.

* Validate 'user_date' parameter, and prevent compile-time bug

The user_date argument wasn't being validated, and there
was a possibility of a compile-time value being used for
initialization; I moved the calculation of initialization-
time date calculation into __init__ and used None as
the "omitted argument" sentinel.

I also used datetime.strptime to validate that the provided
argument is valid.

* Keep the 'user_date' as a string

This is to maintain compatibility with other lines that
make use of user_date

* Moved non-variant initializations to class definition

These attributes do not vary in their initialization depending
on the value of the input to __init__, so I moved them to the
class definition, thus their initialization happens only once.
They are now class attributes, and are inherited by any and all
instances of this class.

* Clarified that __str__ doesn't print anything

__str__ returns a printable representation, but doesn't actually
print anything, so I clarified the comment a bit.  I also removed
the possibility that it would change the internal state of the
object.

* Removed unneeded __call__ method

The object itself was behaving as an alternate constructor,
allowing objects to be callable (classes are always callable);
added in its place a class method 'today()' that instantiates
a new object of the class passing the current date.

* Moved today() to right before the __init__ method

This gives much clarity about what happens when calling
it (it creates a new object and calls __init__ on it).

* Removed redundant optional argument handling in __init__

Since .today() is now available to provide the current date's
timestamp, it is not necessary, and in fact redundant, to have
the same code in __init__ performing the same operations; that
code path has been removed, __init__ now always requires
a timestamp.

* Don't return an error value, raise an exception

Instead of returning an error value, raise a ValueError
indicating that the variant requested is unknown.

* Move this assignment until after validation

This is a little bit of lazy initialization; don't perform any
actions that might end up being thrown away in case of the
exception being raised; instead, move after the validation.

* Test for only one condition instead of two

Instead of comparing against two values and then using one or another
after testing against it also, just check for the one different case
("Short") and use "Long" if it's anything else.

* Eliminated use of a temporary variable

* Rephrased the error message to not suggest this is an interactive tool

This is an API, so the message should be about what it expected to
find, and not what the end user should do.

* Try not to mix time.* with datetime.* objects or methods

* Use f-strings where appropriate

Since we're already using them in many other places in the
module.

* Extract the day of year directly from the datetime object

The datetime object embeds a "time tuple" which in one of its
fields contains the day of the year; instead of calculating it
through a timedelta, just grab the right field.

* Assign the int() converted year to itself

Since the converted year wasn't stored, if it was a numerical
string it would pass the conversion but would blow up when
compared against or subtracted from. :-/

* Separate except clauses for the different types of errors we can catch

Every different error type returns some context about the error that
caused it; also, never "swallow" errors silently; this function originally
returned None if there was an error, now it'll re-raise the same
exception, or it can be re-cast as a custom error type.

Also, execution does not proceed beyond the point when an error was
detected, simplifying the definition's structure.

* Eliminate one of the else: clause by negating the condition

This way, the nested raise() statement remains, but because else:
is not required after a successful raise(), we can simplify the
code.

* Make the exception message terse

By making the exception message terser we can shorten the statement

* Use the methods, don't parse strings if not necessary

* Give startup() a docstring

* Omitted various superfluous else: clauses

After return or raise, there is no need for an else clause to handle
the alternate condition.

Co-authored-by: R.M.Beristain <64236102+RMBeristain@users.noreply.github.com>

* Improvements to README

Fixed some typos and cleaned up a bit of the descriptions.

* Fixed typo in module docstring

Also, Black did its thing and I removed the compile-time-bug-demo sub folder to avoid user confusion.

* Fixed  subtle bug in `get_uniday` that  broke weekday name regularity

-  Moved weekdays list to class property
- Also added note to docstring to clarify that all weeks must start  on Firstday and end on Sixthday.
- Sneaked in The Return of The Side-effect ;) I felt that not being able to instantiate without parameters changed the contract too much.
-  Updated  callable test to reflect the above side effect
- Fixed  references to InvalidUnifiedDateValue to reflect new more descriptive name

* Updated test to use `.today` method instead of passing keyword.

* Updated test to expect ValueError instead of string.

* `format_date` raises exception for missing date + removed  'Today' parameter from `unify`

- format_date: improved docstring  and raise InvalidUnifiedDateValue when unified date property is None.
- updated test to expect InvalidUnifiedDateValue exception instead of a string.
- unify:  no longer accepts "Today" as parameter ; replaced by`.today` method.

* Condensed default value tests into one, using `.today` method.

* reverse_year raises InvalidUnifiedDateValue instead of str, added test

- reverse_year raises InvalidUnifiedDateValue instead of returning a string
- fixed wrong type ind (wasn't optional)
- added test for function
- reverse_unidate_matches: fixed docstring and moved deepcopy import to top

Co-authored-by: rangerelf <gustavo.cordova@gmail.com>